-Transition works now -Character does not stuck anymore -sgfEntity::update(int deltaTime) changed into sgfEntity::update(float deltaTime)

+++ trunk/Simple Game Framework/Simple Game Framework/Entity.cpp
@@ -36,7 +36,7 @@ void sgfEntity::onEvent(SEntityEvent& param)
 	//entityEvent(param);
 }
 
-void sgfEntity::update(int deltaTime)
+void sgfEntity::update(float deltaTime)
 {
 }
